Exception handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -2,12 +2,16 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import java.util.Optional;
 
 //implement the search on database, with dependency ProductRepository
 @Service
@@ -20,10 +24,12 @@ public class ProductService {
     //findById method to search products on database for id
     //only ready operation in database(don't lock in database)
     //first search product through id in database, after handle this product and save one variable, after convert this product in one product dto.
+    //findById have method of optional, that throw exception
+    //orElseThrow() throw exception case don't find object
+    //this Service layer is responsible that throw the own exceptions.
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-            Optional<Product> result = repository.findById(id);
-            Product product = result.get();
+            Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
             ProductDTO dto = new ProductDTO(product);
             return dto;
     }
@@ -52,17 +58,30 @@ public class ProductService {
     //copyDtoToEntity is method to copy dto for entity
     @Transactional
     public ProductDTO update(Long id,  ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e) {
+             throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    //method to delete
+    //method delete
     // is void, don't have return. delete code 204 no content
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     //create method for copy dto to entity
